+++ app.py
@@ -4,7 +4,7 @@ import streamlit as st
 from src.utils import get_args, load_config, set_token, print_message, get_logger, get_secrets
 from src.doc_utils import process_file
 from src.text_preprocessing import get_chunks
-from src.chatbot import bot
+from src.chatbot_v2 import bot
 from src.app_utils import frontend_manager
 
 if __name__ == '__main__':
@@ -39,12 +39,15 @@ if __name__ == '__main__':
         ):
         
         st.session_state['prompting'] = True
-        documents, _ = process_file(sidebar_data['uploaded_file'])
-        
-        chunks = get_chunks(documents)
 
         chatbot = bot(secrets=secrets, config=config, bot_config=sidebar_data, local=False)
         chatbot.initialize()
+        
+        documents, _ = chatbot.process_file(sidebar_data['uploaded_file'])
+        
+        chunks = get_chunks(documents)
+
+
         
         query = prompt_container.chat_input(placeholder="Memento mori!", disabled=False)