+++ src/doc_utils.py
@@ -2,8 +2,20 @@ import os
 from src.utils import print_message
 from PyPDF2 import PdfReader
 from langchain.document_loaders.csv_loader import CSVLoader
-from langchain.docstore.document import Document
+from langchain.text_splitter import CharacterTextSplitter
+# from langchain.docstore.document import Document
 import tempfile
+
+
+from pydantic import Field
+class Document:
+    def __init__(self, page_content: str, metadata: dict = Field(default_factory=dict)) -> None:
+        """Class for storing a piece of text and associated metadata."""
+        """String text."""
+        self.page_content = page_content
+        """Dictionary of metadata."""
+        self.metadata = metadata
+        
 
 def process_pdf(file, documents):
     pdfReader = PdfReader(file)
@@ -13,7 +25,8 @@ def process_pdf(file, documents):
     for i, page in enumerate(pdfReader.pages):
         documents.append(Document(page_content=page.extract_text(),
                                 metadata={'title': file, 
-                                            'page_number': i}))
+                                            'page_number': i,
+                                            'file_name': file}))
     return documents
 
 def process_text(file, documents):
@@ -21,7 +34,8 @@ def process_text(file, documents):
     for line in file:
         text +=  str(line.decode())
     documents.append(Document(page_content=text,
-                            metadata={'title': file}))
+                            metadata={'title': file,
+                                      'file_name': file}))
     return documents
 
 def process_csv(file, documents):
@@ -37,82 +51,88 @@ def process_csv(file, documents):
         data = loader.load()
         return data
 
+def get_chunks(documents):
+        
+    char_splitter = CharacterTextSplitter(chunk_size=1000, chunk_overlap=200, separator=' ')
+    chunks = char_splitter.split_documents(documents)
+
+    for i, chunk in enumerate(chunks):
+        if len(chunk.page_content)>1000:
+            chunks[i].page_content = chunk.page_content[:1000]
+    return chunks
+
 
 def process_file(uploaded_file):
-    
-    # message = 'Processing your files, please wait. This will take a few mins.'
-    # print_message(message, st=st)
-    
-    files = []
+    """
+    1. Check for new files
+    2. Fow new files - 
+        1. Read the files
+        2. Create documents
+        3. Create chunks
+        4. Return chunks
+    """
     documents = []
-    for file in uploaded_file:
-        
-        # files.append(file.name)
-        # message = f'Processing document - {file.name} which is a {file.type} file. Please wait...'
-        # print_message(message, st=st)
-        
-        if file[-4:]=='.pdf': 
-            documents = process_pdf(file, documents)
-        
-        if file[-4:]=='.txt':
-            documents = process_text(file, documents)
-        
-        # if file.type=='text/csv':
-        #     documents = process_csv(file, st, documents)
+    if uploaded_file[-4:]=='.pdf': 
+        documents = process_pdf(uploaded_file, documents)
+    
+    if uploaded_file[-4:]=='.txt':
+        documents = process_text(uploaded_file, documents)
 
-    # new_files = check_for_new_files(files, st)
-    new_files = False
-    return documents, new_files
+    chunks = get_chunks(documents)
+    return chunks
 
 
-def check_for_files(st):
-    if not os.path.exists('app_database'):
-        os.mkdir('app_database')
-        message = 'Created new folder - app_database'
-        print_message(message, st=st)
-    if not os.path.exists('app_database/db'):
-        os.mkdir('app_database/db')
-        message = 'Created new folder - app_database/db'
-        print_message(message, st=st)
-    if not os.path.exists('app_database/files.txt'):
-        with open('app_database/files.txt','w') as f:
-            message = 'Created new file - files.txt'
-            print_message(message, st=st)
-            f.close()
-    if not os.path.exists('app_database/result.txt'):
-        with open('app_database/result.txt','w') as f:
-            message = 'Created new file - result.txt'
-            print_message(message, st=st)
-            f.close()
-    if not os.path.exists('app_database/query_ids.txt'):
-        with open('app_database/query_ids.txt','w') as f:
-            message = 'Created new file - query_ids.txt'
-            print_message(message, st=st)
-            f.close()
-
-def check_for_new_files(files, st):
-    check_for_files(st)
-    files = list(set(files))
-    old_files = read_old_file_names()
-    if old_files==files:
-        message = 'No new files uploaded. Using the existing database.'
-        print_message(message, st=st)
-        return False
-    else:
-        write_new_file_names(files)
-        return True
 
 
-def read_old_file_names():
-    with open('app_database/files.txt','r') as f:
-        files = f.readlines()
-        f.close()
-    files = [file.strip() for file in files]
-    return files
 
-def write_new_file_names(files):
-    with open('app_database/files.txt','w') as f:
-        for item in files:
-            f.write(item+"\n")
-        f.close()
-    return None
+# def check_for_files(st):
+#     if not os.path.exists('app_database'):
+#         os.mkdir('app_database')
+#         message = 'Created new folder - app_database'
+#         print_message(message, st=st)
+#     if not os.path.exists('app_database/db'):
+#         os.mkdir('app_database/db')
+#         message = 'Created new folder - app_database/db'
+#         print_message(message, st=st)
+#     if not os.path.exists('app_database/files.txt'):
+#         with open('app_database/files.txt','w') as f:
+#             message = 'Created new file - files.txt'
+#             print_message(message, st=st)
+#             f.close()
+#     if not os.path.exists('app_database/result.txt'):
+#         with open('app_database/result.txt','w') as f:
+#             message = 'Created new file - result.txt'
+#             print_message(message, st=st)
+#             f.close()
+#     if not os.path.exists('app_database/query_ids.txt'):
+#         with open('app_database/query_ids.txt','w') as f:
+#             message = 'Created new file - query_ids.txt'
+#             print_message(message, st=st)
+#             f.close()
+
+# def check_for_new_files(files, st):
+#     check_for_files(st)
+#     files = list(set(files))
+#     old_files = read_old_file_names()
+#     if old_files==files:
+#         message = 'No new files uploaded. Using the existing database.'
+#         print_message(message, st=st)
+#         return False
+#     else:
+#         write_new_file_names(files)
+#         return True
+
+
+# def read_old_file_names():
+#     with open('app_database/files.txt','r') as f:
+#         files = f.readlines()
+#         f.close()
+#     files = [file.strip() for file in files]
+#     return files
+
+# def write_new_file_names(files):
+#     with open('app_database/files.txt','w') as f:
+#         for item in files:
+#             f.write(item+"\n")
+#         f.close()
+#     return None